% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid.R
\name{get_centroid}
\alias{get_centroid}
\title{Get the centroids of preciencts or wards}
\usage{
get_centroid(Ward = FALSE, Precinct = FALSE)
}
\arguments{
\item{Ward}{Input True for ward list}

\item{Precinct}{Input True for precinct list}
}
\description{
Get the centroids of either the precincts or wards of DC
}
\examples{
precinctCentroids <- get_centroid(Ward = TRUE)
wardCentroids <- get_centroid(Precinct = TRUE)
}
