% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{get_Precinct}
\alias{get_Precinct}
\title{Get the dataframe or shapefile of precincts in DC}
\usage{
get_Precinct(dataframe = TRUE)
}
\arguments{
\item{dataframe}{Put TRUE for a dataframe, FALSE for a shapeFile}
}
\description{
Get the dataframe or shapefile of precincts in DC
}
\examples{
bingo <- get_Precinct(dataframe = TRUE)
bongo <- get_Precinct(dataframe = FALSE)
}
