% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Ward.R
\name{get_Ward}
\alias{get_Ward}
\title{get the Ward shape or Dataframe}
\usage{
get_Ward(dataframe = TRUE)
}
\arguments{
\item{dataframe}{Either TRUE or FALSE}
}
\description{
get the Ward shape or Dataframe
}
\examples{
get_Ward(dataframe = TRUE)
get_Ward(dataframe = FALSE)
}
